RavenDB-18009 added the ability to proxy /databases endpoint
