+++ src/Raven.Server/Web/System/DatabasesHandler.cs
@@ -3,21 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
-using CsvHelper;
-using Raven.Client.Documents.Indexes;
-using Raven.Client.Documents.Operations.Configuration;
-using Raven.Client.Documents.Operations.Replication;
-using Raven.Client.Exceptions;
 using Raven.Client.Exceptions.Database;
-using Raven.Client.Extensions;
 using Raven.Client.Http;
 using Raven.Client.ServerWide;
 using Raven.Client.ServerWide.Commands;
 using Raven.Client.ServerWide.Operations;
 using Raven.Client.Util;
-using Raven.Server.Config;
-using Raven.Server.Documents;
-using Raven.Server.Documents.PeriodicBackup;
 using Raven.Server.Extensions;
 using Raven.Server.NotificationCenter.Notifications;
 using Raven.Server.Routing;
@@ -26,63 +17,20 @@ using Raven.Server.ServerWide.Commands;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.Smuggler.Migration;
 using Raven.Server.Utils;
+using Raven.Server.Web.System.Processors.Databases;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
 using Sparrow.Logging;
-using Constants = Raven.Client.Constants;
 
 namespace Raven.Server.Web.System
 {
     public sealed class DatabasesHandler : RequestHandler
     {
-        private static readonly Logger Logger = LoggingSource.Instance.GetLogger<DatabasesHandler>("Server");
-
         [RavenAction("/databases", "GET", AuthorizationStatus.ValidUser, EndpointType.Read)]
         public async Task Databases()
         {
-            // if Studio requested information about single resource - handle it
-            var dbName = GetStringQueryString("name", false);
-            if (dbName != null)
-            {
-                await DbInfo(dbName);
-                return;
-            }
-
-            var namesOnly = GetBoolValueQueryString("namesOnly", required: false) ?? false;
-
-            using (ServerStore.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
-            {
-                context.OpenReadTransaction();
-                await using (var writer = new AsyncBlittableJsonTextWriter(context, ResponseBodyStream()))
-                {
-                    writer.WriteStartObject();
-
-                    var items = ServerStore.Cluster.GetAllRawDatabases(context, GetStart(), GetPageSize());
-
-                    var allowedDbs = await GetAllowedDbsAsync(null, requireAdmin: false, requireWrite: false);
-
-                    if (allowedDbs.HasAccess == false)
-                        return;
-
-                    if (allowedDbs.AuthorizedDatabases != null)
-                    {
-                        items = items.Where(item => allowedDbs.AuthorizedDatabases.ContainsKey(item.DatabaseName));
-                    }
-
-                    writer.WriteArray(context, nameof(DatabasesInfo.Databases), items.SelectMany(x => x.AsShardsOrNormal()), (w, c, dbDoc) =>
-                    {
-                        var databaseName = dbDoc.DatabaseName;
-                        if (namesOnly)
-                        {
-                            w.WriteString(databaseName);
-                            return;
-                        }
-
-                        WriteDatabaseInfo(databaseName, dbDoc, context, w);
-                    });
-                    writer.WriteEndObject();
-                }
-            }
+            using (var processor = new DatabasesHandlerProcessorForGet(this))
+                await processor.ExecuteAsync();
         }
 
         [RavenAction("/admin/databases/topology/modify", "POST", AuthorizationStatus.Operator)]
@@ -380,293 +328,6 @@ namespace Raven.Server.Web.System
                 url = clusterTopology.GetUrlFromTag(tag);
 
             return url;
-        }
-
-        private async Task DbInfo(string dbName)
-        {
-            using (ServerStore.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
-            {
-                context.OpenReadTransaction();
-                await using (var writer = new AsyncBlittableJsonTextWriter(context, ResponseBodyStream()))
-                {
-                    var dbRecord = ServerStore.Cluster.ReadRawDatabaseRecord(context, dbName, out long _);
-                    
-                    if (dbRecord == null)
-                    {
-                        HttpContext.Response.Headers.Remove(Constants.Headers.ContentType);
-                        HttpContext.Response.StatusCode = (int)HttpStatusCode.NotFound;
-                        HttpContext.Response.Headers["Database-Missing"] = dbName;
-                        return;
-                    }
-
-                    WriteDatabaseInfo(dbName, dbRecord, context, writer);
-                    
-                }
-            }
-        }
-
-        private void WriteDatabaseInfo(string databaseName, RawDatabaseRecord rawDatabaseRecord,
-            TransactionOperationContext context, AbstractBlittableJsonTextWriter writer)
-        {
-            var nodesTopology = new NodesTopology();
-
-            try
-            {
-                var online = ServerStore.DatabasesLandlord.DatabasesCache.TryGetValue(databaseName, out Task<DocumentDatabase> dbTask) &&
-                             dbTask != null &&
-                             dbTask.IsCompleted;
-                
-                var topology = rawDatabaseRecord.IsSharded ? rawDatabaseRecord.Sharding.Orchestrator.Topology : rawDatabaseRecord.Topology;
-
-                var statuses = ServerStore.GetNodesStatuses();
-                if (topology != null)
-                {
-                    nodesTopology.PriorityOrder = topology.PriorityOrder;
-
-                    var clusterTopology = ServerStore.GetClusterTopology(context);
-                    clusterTopology.ReplaceCurrentNodeUrlWithClientRequestedNodeUrlIfNecessary(ServerStore, HttpContext);
-
-                    foreach (var member in topology.Members)
-                    {
-                        if (rawDatabaseRecord.DeletionInProgress != null && rawDatabaseRecord.DeletionInProgress.ContainsKey(member))
-                            continue;
-
-                        var url = clusterTopology.GetUrlFromTag(member);
-                        var node = new InternalReplication
-                        {
-                            Database = databaseName,
-                            NodeTag = member,
-                            Url = url
-                        };
-                        nodesTopology.Members.Add(GetNodeId(node));
-                        SetNodeStatus(topology, member, nodesTopology, statuses);
-                    }
-
-                    foreach (var promotable in topology.Promotables)
-                    {
-                        if (rawDatabaseRecord.DeletionInProgress != null && rawDatabaseRecord.DeletionInProgress.ContainsKey(promotable))
-                            continue;
-
-                        topology.PredefinedMentors.TryGetValue(promotable, out var mentorCandidate);
-                        var node = GetNode(databaseName, clusterTopology, promotable, mentorCandidate, out var promotableTask);
-                        var mentor = topology.WhoseTaskIsIt(ServerStore.Engine.CurrentState, promotableTask, null);
-                        nodesTopology.Promotables.Add(GetNodeId(node, mentor));
-                        SetNodeStatus(topology, promotable, nodesTopology, statuses);
-                    }
-
-                    foreach (var rehab in topology.Rehabs)
-                    {
-                        if (rawDatabaseRecord.DeletionInProgress != null && rawDatabaseRecord.DeletionInProgress.ContainsKey(rehab))
-                            continue;
-
-                        var node = GetNode(databaseName, clusterTopology, rehab, null, out var promotableTask);
-                        var mentor = topology.WhoseTaskIsIt(ServerStore.Engine.CurrentState, promotableTask, null);
-                        nodesTopology.Rehabs.Add(GetNodeId(node, mentor));
-                        SetNodeStatus(topology, rehab, nodesTopology, statuses);
-                    }
-                }
-
-                // Check for exceptions
-                if (dbTask != null && dbTask.IsFaulted)
-                {
-                    var exception = dbTask.Exception.ExtractSingleInnerException();
-                    WriteFaultedDatabaseInfo(databaseName, nodesTopology, exception, context, writer);
-                    return;
-                }
-
-                var db = online ? dbTask.Result : null;
-
-                var indexingStatus = db?.IndexStore?.Status;
-                if (indexingStatus == null)
-                {
-                    // Looking for disabled indexing flag inside the database settings for offline database status
-                    if (rawDatabaseRecord.Settings.TryGetValue(RavenConfiguration.GetKey(x => x.Indexing.Disabled), out var val) &&
-                        bool.TryParse(val, out var indexingDisabled) && indexingDisabled)
-                        indexingStatus = IndexRunningStatus.Disabled;
-                }
-
-                var disabled = rawDatabaseRecord.IsDisabled;
-                var lockMode = rawDatabaseRecord.LockMode;
-
-                var studioEnvironment = StudioConfiguration.StudioEnvironment.None;
-                if (rawDatabaseRecord.StudioConfiguration != null && !rawDatabaseRecord.StudioConfiguration.Disabled)
-                {
-                    studioEnvironment = rawDatabaseRecord.StudioConfiguration.Environment;
-                }
-
-                if (online == false)
-                {
-                    // if state of database is found in the cache we can continue
-                    if (ServerStore.DatabaseInfoCache.TryGet(databaseName, databaseInfoJson =>
-                    {
-
-                        var periodicBackups = new List<PeriodicBackup>();
-
-                        foreach (var periodicBackupConfiguration in rawDatabaseRecord.PeriodicBackupsConfiguration)
-                        {
-                            periodicBackups.Add(new PeriodicBackup
-                            {
-                                Configuration = periodicBackupConfiguration,
-                                BackupStatus = BackupUtils.GetBackupStatusFromCluster(ServerStore, context, databaseName, periodicBackupConfiguration.TaskId)
-                            });
-                        }
-
-                        databaseInfoJson.Modifications = new DynamicJsonValue(databaseInfoJson)
-                        {
-                            [nameof(DatabaseInfo.Disabled)] = disabled,
-                            [nameof(DatabaseInfo.LockMode)] = lockMode,
-                            [nameof(DatabaseInfo.IndexingStatus)] = indexingStatus,
-                            [nameof(DatabaseInfo.NodesTopology)] = nodesTopology.ToJson(),
-                            [nameof(DatabaseInfo.DeletionInProgress)] = DynamicJsonValue.Convert(rawDatabaseRecord.DeletionInProgress),
-                            [nameof(DatabaseInfo.Environment)] = studioEnvironment,
-                            [nameof(DatabaseInfo.BackupInfo)] = BackupUtils.GetBackupInfo(
-                                new BackupUtils.BackupInfoParameters()
-                                {
-                                    ServerStore = ServerStore,
-                                    PeriodicBackups = periodicBackups,
-                                    DatabaseName = databaseName,
-                                    Context = context
-                                })
-                        };
-
-                        context.Write(writer, databaseInfoJson);
-                    }))
-                    {
-                        return;
-                    }
-
-                    // we won't find it if it is a new database or after a dirty shutdown,
-                    // so just report empty values then
-                }
-
-                var size = db?.GetSizeOnDisk() ?? (new Size(0), new Size(0));
-
-                var databaseInfo = new DatabaseInfo
-                {
-                    Name = databaseName,
-                    Disabled = disabled,
-                    LockMode = lockMode,
-                    TotalSize = size.Data,
-                    TempBuffersSize = size.TempBuffers,
-
-                    IsAdmin = true,
-                    IsEncrypted = rawDatabaseRecord.IsEncrypted,
-                    UpTime = online ? (TimeSpan?)GetUptime(db) : null,
-                    BackupInfo = GetBackupInfo(db, context),
-
-                    Alerts = db?.NotificationCenter.GetAlertCount() ?? 0,
-                    PerformanceHints = db?.NotificationCenter.GetPerformanceHintCount() ?? 0,
-                    RejectClients = false,
-                    LoadError = null,
-                    IndexingErrors = db?.IndexStore?.GetIndexes()?.Sum(index => index.GetErrorCount()) ?? 0,
-
-                    DocumentsCount = db?.DocumentsStorage.GetNumberOfDocuments() ?? 0,
-                    HasRevisionsConfiguration = db?.DocumentsStorage.RevisionsStorage.Configuration != null,
-                    HasExpirationConfiguration = (db?.ExpiredDocumentsCleaner?.ExpirationConfiguration?.Disabled ?? true) == false,
-                    HasRefreshConfiguration = (db?.ExpiredDocumentsCleaner?.RefreshConfiguration?.Disabled ?? true) == false,
-                    IndexesCount = db?.IndexStore?.GetIndexes()?.Count() ?? 0,
-                    IndexingStatus = indexingStatus ?? IndexRunningStatus.Running,
-                    Environment = studioEnvironment,
-
-                    NodesTopology = nodesTopology,
-                    ReplicationFactor = topology?.ReplicationFactor ?? -1,
-                    DynamicNodesDistribution = topology?.DynamicNodesDistribution ?? false,
-                    DeletionInProgress = rawDatabaseRecord.DeletionInProgress
-                };
-
-                var doc = databaseInfo.ToJson();
-                context.Write(writer, doc);
-            }
-            catch (Exception e)
-            {
-                if (Logger.IsInfoEnabled)
-                    Logger.Info($"Failed to get database info for: {databaseName}", e);
-
-                WriteFaultedDatabaseInfo(databaseName, nodesTopology, e, context, writer);
-            }
-        }
-
-        private static void SetNodeStatus(
-            DatabaseTopology topology,
-            string nodeTag,
-            NodesTopology nodesTopology,
-            Dictionary<string, NodeStatus> nodeStatuses)
-        {
-            var nodeStatus = new DatabaseGroupNodeStatus
-            {
-                LastStatus = DatabasePromotionStatus.Ok
-            };
-            if (topology.PromotablesStatus.TryGetValue(nodeTag, out var status))
-            {
-                nodeStatus.LastStatus = status;
-            }
-            if (topology.DemotionReasons.TryGetValue(nodeTag, out var reason))
-            {
-                nodeStatus.LastError = reason;
-            }
-
-            if (nodeStatus.LastStatus == DatabasePromotionStatus.Ok &&
-                nodeStatuses.TryGetValue(nodeTag, out var serverNodeStatus) &&
-                serverNodeStatus.Connected == false)
-            {
-                nodeStatus.LastError = serverNodeStatus.ErrorDetails;
-                nodeStatus.LastStatus = DatabasePromotionStatus.NotResponding;
-            }
-
-            nodesTopology.Status[nodeTag] = nodeStatus;
-        }
-
-        private static InternalReplication GetNode(string databaseName, ClusterTopology clusterTopology, string rehab, string mentor, out PromotableTask promotableTask)
-        {
-            var url = clusterTopology.GetUrlFromTag(rehab);
-            var node = new InternalReplication
-            {
-                Database = databaseName,
-                NodeTag = rehab,
-                Url = url
-            };
-            promotableTask = new PromotableTask(rehab, url, databaseName, mentor);
-            return node;
-        }
-
-        private void WriteFaultedDatabaseInfo(string databaseName,
-            NodesTopology nodesTopology,
-            Exception exception,
-            JsonOperationContext context,
-            AbstractBlittableJsonTextWriter writer)
-        {
-            var doc = new DynamicJsonValue
-            {
-                [nameof(DatabaseInfo.Name)] = databaseName,
-                [nameof(DatabaseInfo.NodesTopology)] = nodesTopology,
-                [nameof(DatabaseInfo.LoadError)] = exception.Message
-            };
-
-            context.Write(writer, doc);
-        }
-
-        private static BackupInfo GetBackupInfo(DocumentDatabase db, TransactionOperationContext context)
-        {
-            var periodicBackupRunner = db?.PeriodicBackupRunner;
-            var results = periodicBackupRunner?.GetBackupInfo(context);
-            return results;
-        }
-
-        private static TimeSpan GetUptime(DocumentDatabase db)
-        {
-            return SystemTime.UtcNow - db.StartTime;
-        }
-
-        private static NodeId GetNodeId(InternalReplication node, string responsible = null)
-        {
-            var nodeId = new NodeId
-            {
-                NodeTag = node.NodeTag,
-                NodeUrl = node.Url,
-                ResponsibleNode = responsible
-            };
-
-            return nodeId;
         }
     }
 }
